split suspensePromise.ts
